se agrega correciones en archivos
