+++ src/main/java/com/sofkau/tasks/automationexcercise/AbrirPaginaInicial.java
@@ -1,16 +1,23 @@
 package com.sofkau.tasks.automationexcercise;
 
-import com.sofkau.ui.PaginaInicial;
 import net.serenitybdd.screenplay.Actor;
 import net.serenitybdd.screenplay.Task;
 import net.serenitybdd.screenplay.actions.Open;
 
 public class AbrirPaginaInicial implements Task {
-    private PaginaInicial paginaInicial;
+
+    private String paginaInicial;
+    public AbrirPaginaInicial conLaPagina(String paginaInicial){
+        this.paginaInicial = paginaInicial;
+        return this;
+    }
     @Override
     public <T extends Actor> void performAs(T actor) {
         actor.attemptsTo(
-                Open.browserOn(paginaInicial)
+                Open.url(paginaInicial)
         );
+    }
+    public static AbrirPaginaInicial abrirPaginaInicial(){
+        return new AbrirPaginaInicial();
     }
 }
