--- src/main/java/com/sofkau/ui/PaginaInicial.java
@@ -1,6 +1,0 @@
-package com.sofkau.ui;
-
-import net.thucydides.core.pages.PageObject;
-
-public class PaginaInicial extends PageObject {
-}
